refactor(file-structure): shifting components under individual _commponents folders
