+++ app/(auth)/signup/_components/signup-form.tsx
@@ -10,8 +10,8 @@ import { useAuthState } from '@/hooks/useAuthState';
 // LIBRARIES
 import { SubmitHandler, useForm } from 'react-hook-form';
 // COMPONENTS
-import InputField from '@/components/InputField';
-import { Button, buttonVariants } from '@/components/ui/Button';
+import { InputField } from '@/components/input-field';
+import { Button, buttonVariants } from '@/components/ui/button';
 // CONFIG
 import { signUpInputs } from '@/config/forms';
