+++ app/(auth)/signup/page.tsx
@@ -2,9 +2,9 @@
 import { type Metadata } from 'next';
 // COMPONENTS
 import { SignUpForm } from './_components/signup-form';
-import Header from '@/components/Header';
-import Title, { titleVariants } from '@/components/ui/Title';
-import Description, { descriptionVariants } from '@/components/ui/Description';
+import { Header } from "@/components/ui/header";
+import { Title, titleVariants } from '@/components/ui/title';
+import { Description, descriptionVariants } from "@/components/ui/description";
 // UTILS
 import { cn } from '@/utils/helpers';
