--- components/Avatar/types.ts
@@ -1,6 +1,0 @@
-export interface AvatarProps {
-  src: string | undefined | null;
-  width: number;
-  height: number;
-  altText?: string;
-}
