--- components/Buttons/index.tsx
@@ -1,9 +1,0 @@
-import { CheckoutButton } from "./CheckoutButton";
-import { ManageSubscriptionButton } from "../../app/dashboard/billing/_components/manage-subscription-button";
-import { ToggleAutoPlayButton } from "./ToggleAutoPlayButton";
-
-export { 
-  CheckoutButton, 
-  ManageSubscriptionButton, 
-  ToggleAutoPlayButton
-};
