--- components/InputField/types.ts
@@ -1,15 +1,0 @@
-import { UseFormRegister } from "react-hook-form";
-export interface InputFieldProps {
-  name: string;
-  label?: string;
-  type?: string;
-  placeholder?: string;
-  register: UseFormRegister<any>;
-  validationSchema?: {
-    required?: string;
-    minLength?: { value: number; message: string };
-    maxLength?: { value: number; message: string };
-    pattern?: { value: RegExp; message: string };
-  }
-  error?: string; 
-}
