+++ components/auth-check.tsx
@@ -3,7 +3,7 @@
 // HOOKS
 import { useAuthState } from '@/hooks/useAuthState';
 
-export default function AuthCheck(props: any) {
+export function AuthCheck(props: any) {
   const { user } = useAuthState();
 
   return user ? <>{props.children}</> : props.fallback || null;
