+++ components/input-field.tsx
@@ -1,8 +1,24 @@
 "use client"
 
-import { InputFieldProps } from './types';
+// PACKAGES
+import { UseFormRegister } from "react-hook-form";
 
-export default function InputField({
+export interface InputFieldProps {
+  name: string;
+  label?: string;
+  type?: string;
+  placeholder?: string;
+  register: UseFormRegister<any>;
+  validationSchema?: {
+    required?: string;
+    minLength?: { value: number; message: string };
+    maxLength?: { value: number; message: string };
+    pattern?: { value: RegExp; message: string };
+  }
+  error?: string; 
+}
+
+export function InputField({
   name,
   label,
   register,
