+++ components/subscription-check.tsx
@@ -3,7 +3,7 @@
 // HOOKS
 import { useIsSubscriptionValid } from '@/hooks/useIsSubscriptionValid';
 
-export default function Subscriptioncheck(props: any) {
+export function Subscriptioncheck(props: any) {
   const isUserPro = useIsSubscriptionValid();
 
   return !isUserPro ? <>{props.children}</> : props.fallback || null;
