+++ components/ui/spinner.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 export default function Spinner() {
   return (
     <div
