+++ types/index.ts
@@ -85,3 +85,22 @@ export interface FetchFromApiOptions {
   body?: any;
   cache?: string;
 }
+
+export interface AvatarProps {
+  src: string | undefined | null;
+  width: number;
+  height: number;
+  altText?: string;
+}
+
+export interface ImageWithFallbackProps {
+  src: string | null | undefined;
+  fallbackSrc: string;
+  width?: number;
+  height?: number;
+  altText?: string;
+  layout?: "fixed" | "intrinsic" | "responsive" | "fill";
+  objectFit?: "fill" | "contain" | "cover" | "none" | "scale-down";
+  sizes?: string;
+  style?: React.CSSProperties;
+}
